fix: add isActive flag to undo redo in guide. refactor: add custom hook to Undo component
