+++ package.json
@@ -10,8 +10,8 @@
     "@types/node": "^16.18.11",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
-    "i18next": "^22.4.10",
     "firebase": "^9.17.1",
+    "i18next": "^22.4.10",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^12.1.5",
